requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
pygments
coreapi
django-rest-elasticsearch